R - move DuckInode into a separate header
